Refactor trace store manifest budget helpers
